STATBANK-453 Add react only app (#1)
* Add a new react app to the Swedish poc

Since we are starting on a new project with statisticssweden in January,
we need a place to both learn React, and test what we want to do. This
commit adds a react application to the POC project repository that
Sweden made.

This is only a skeleton, set up with the basics we need at the start.
It's missing things like language support, and tests have not been
written beyond what is created with Nx.